checkpoint: pending pipeline work (train_head head + refinements)
Snapshot of uncommitted non-demo work before applying concern-remediation
fixes. Includes: train_head.py read-out head, Makefile head/head-demo targets,
dry_run.py section 8 (head signal+null+shuffle controls), publish_results.py
verdict refinement, honest_corr ceiling-docstring honesty fix, and
backlog/prereg-head/product docs. Demo/ left uncommitted (live edit session).

+++ batch_extract.py
@@ -220,7 +220,7 @@ def main():
                 events = build_events(vp)
                 preds, _segments = model.predict(events=events)
                 preds = np.asarray(preds, float)
-                np.save(npy, preds)
+                np.save(npy, preds.astype(np.float32))  # float32 halves disk/IO; z-scored BOLD needs no more (matches Colab Cell 4)
             if not printed_scale:
                 describe_preds(preds, tag=vid)   # confirm z-scored/signed scale ONCE
                 printed_scale = True
